# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/security/session/refresh": {
      "post": {
        "operationId": "SessionController_refresh",
        "summary": "Renueva la sesión rotando el refresh token",
        "description": "Lee la cookie `auth.refresh`, revoca la sesión anterior y devuelve una nueva cookie de refresh. Un refresh token ya rotado da 401.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión renovada (cookie `auth.refresh` nueva).",
            "headers": {
              "Set-Cookie": {
                "description": "auth.refresh renovada.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Sesión renovada"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh token ausente, inválido, expirado o revocado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Sesión"
        ],
        "security": [
          {
            "refresh": []
          }
        ]
      }
    },
    "/security/logout": {
      "get": {
        "operationId": "SessionController_logout",
        "summary": "Cierra la sesión",
        "description": "Revoca la sesión en base de datos y caché, destruye la sesión HTTP y borra las cookies `auth.refresh` y `auth.session`. Nota: cambia estado con GET (pendiente migrar a POST).",
        "deprecated": false,
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión cerrada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Logout exitoso"
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Sesión"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/registro/check/{field}": {
      "get": {
        "operationId": "RegistroController_getRegistro",
        "summary": "Comprueba si un campo (p. ej. username o correo) está disponible",
        "parameters": [
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "example": "username",
              "type": "string"
            }
          },
          {
            "name": "value",
            "required": true,
            "in": "query",
            "description": "Valor a comprobar.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disponibilidad del valor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Registro"
        ]
      }
    },
    "/registro": {
      "post": {
        "operationId": "RegistroController_createRegistro",
        "summary": "Registra un usuario; envía un código OTP al correo para verificarlo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro creado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o usuario/correo ya registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Registro"
        ]
      }
    },
    "/registro/verificar-email": {
      "post": {
        "operationId": "RegistroController_verificarEmail",
        "summary": "Verifica el correo con el código OTP de 6 dígitos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificarEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correo verificado; ya se puede iniciar sesión."
          },
          "400": {
            "description": "Código inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Registro"
        ]
      }
    },
    "/registro/resend-otp": {
      "post": {
        "operationId": "RegistroController_resendOtp",
        "summary": "Reenvía el código OTP (respuesta genérica: no revela si el correo existe)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud aceptada."
          }
        },
        "tags": [
          "Registro"
        ]
      }
    },
...
  },
  "info": {
    "title": "ms-identity",
    "description": "Identidad y autenticación de SEIS App: Authorization Code + PKCE, ciclo de sesión, recuperación de contraseña, registro y perfil de usuario.\n\n**Flujo de login**: `POST /security/authorize` (credenciales + code_challenge) → código → `POST /security/token` (código + code_verifier) → cookies `auth.session` y `auth.refresh`. Las rutas de perfil usan el JWT de acceso como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth.session"
      },
      "refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth.refresh"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Usuario o contraseña incorrectos"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "RefreshSessionRequestDto": {
        "type": "object",
        "properties": {
          "typeDevice": {
            "type": "string",
            "example": "WEB",
            "description": "Solo informativo (se usa en logs)."
          }
        }
      },
...
      "FormRegisterDto": {
        "type": "object",
        "properties": {
          "rol": {
            "type": "string",
            "enum": [
              "USUARIO_STANDARD",
              "CEDENTE",
              "CEDENTE_ADM",
              "EJECUTIVO",
              "EJECUTIVO_ADM"
            ],
            "description": "Rol inicial solicitado (el rol efectivo se asigna al verificar el correo)."
          },
          "email": {
            "type": "string",
            "example": "ana@correo.cl"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "example": "ana.perez"
          },
          "nombres": {
            "type": "string",
            "minLength": 2
          },
          "apellidoPaterno": {
            "type": "string",
            "minLength": 2
          },
          "apellidoMaterno": {
            "type": "string",
            "minLength": 2
          },
          "telefono": {
            "type": "string",
            "minLength": 7,
            "maxLength": 20,
            "example": "+56912345678"
          },
          "tipoDocumento": {
            "type": "string",
            "enum": [
              "DNI",
              "RUT",
              "PASAPORTE"
            ]
          },
          "numeroDocumento": {
            "type": "string",
            "example": "12345678-9"
          },
          "pais": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "example": "CL",
            "description": "Código ISO de 2 letras."
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date",
            "example": "1990-05-31"
          },
          "direccion": {
            "type": "string",
            "minLength": 5
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "format": "password"
          }
        },
        "required": [
          "rol",
          "email",
          "username",
          "nombres",
          "apellidoPaterno",
          "apellidoMaterno",
          "telefono",
          "tipoDocumento",
          "numeroDocumento",
          "pais",
          "fechaNacimiento",
          "direccion",
          "password"
        ]
      },
      "VerificarEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ana@correo.cl"
          },
          "otp": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "example": "123456"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResendOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ana@correo.cl"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}